Reorganize how Input component is structured
Break apart the Input component from a singular file into a directory of
multiple components / files.
